Add advancements for custom crafting recipes.
It feels so official!
Please note: This system is meant to replicate "coming up with ideas". For recipes that require multiple items, you must have both in your inventory to "think of" the idea to make the resulting item. Using or droping one of the required items before getting the rest will not give you the recipe. You do not need the correct ammount of items, just one of each type required.